refactor(ngcc): move command line option parsing to its own file
